Fixed KHI not to send requests to a project while user is writing project ID on the form (#265)
* [bugfix :bug:] Fixed KHI not to send requests to a project during the user is writing project ID on the form

The validation for the new inspection dialogues are handled in the backend.
Every user change trigger the dryrun request to backend to verify its values are right or not. But some fields needs to access Google Cloud API and it's not good to trigger this on every change especially user is modifying the project ID.
This change introduced a new parameter ValidatingTiming on the form parameter and the text form won't trigger the validation request to the backend on every changes when the form field had `onblur` as its validating timing.

* fixed failing backend tests

* Fix PR reviewed points

* Update the enum parameter names: OnChange -> Change, OnBlur -> Blur
* Defined a type representing validation timings on frontend